Fix broken link and grammar issue
The `Contributors Guide` link is broken due to the file extension being all lowercase in the link and the actual file's extension being upper case. For consistency, all file extensions should be lowercase.

Also, there should be an apostrophe after the `s` in `Contributors` but it's more common to simply say "Contribution Guide" so changed it to that.